requirements.txt
pytesseract
opencv-python
numpy
pillow
